requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-vertexai>=2.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
gradio>=4.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pyyaml>=6.0
asyncio-throttle>=1.0.2
